backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Computer Vision & ML
mediapipe>=0.10.0
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=10.0.0

# Configuration & Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
httpx>=0.24.0

# Optional: Advanced features
# torch>=2.0.0  # For SMPL & VITON (heavy install)
# redis>=4.5.0  # For production session management
# psycopg2-binary>=2.9.0  # For PostgreSQL
# pymongo>=4.3.0  # For MongoDB
# celery>=5.2.0  # For background tasks
# python-jose[cryptography]>=3.3.0  # For JWT authentication
# passlib[bcrypt]>=1.7.4  # For password hashing
